# docs/api-spec.yaml
openapi: 3.1.0
info:
  title: MockAgents API
  version: 0.4.0
  description: >
    MockAgents is a platform for mocking AI agent integrations. It exposes a
    Management API (agents, interaction logs, costs, audit, pipelines, config
    validation, and a multi-tenant control plane of tenants + RBAC API keys),
    and Protocol Endpoints that mimic real LLM provider APIs (OpenAI Chat
    Completions and Anthropic Messages).


    The Model Context Protocol mock (`/mcp`, `/mcp/events`, `/mcp/response`) is
    JSON-RPC 2.0 rather than REST and is documented in the project README,
    not in this OpenAPI document.


    **Authentication.** When the server runs with `MOCKAGENTS_MULTI_TENANT=1`,
    every `/api/v1/*` management route requires an API key
    (`Authorization: Bearer <key>` or `X-Api-Key: <key>`) whose role meets the
    route's floor (`viewer < editor < admin < platform`). In single-tenant mode
    the management API is open. The protocol endpoints (`/v1/*`) are always
    unauthenticated — clients send their own provider key, which MockAgents
    ignores.
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: http://localhost:8080
    description: Local development server
...
paths:
...
  /api/v1/tenants:
    get:
      operationId: listTenants
      summary: List tenants
      description: Requires the `platform` role.
      tags: [Tenants]
      responses:
        "200":
          description: Tenants.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Tenant" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
    post:
      operationId: createTenant
      summary: Create a tenant
      description: Requires the `platform` role.
      tags: [Tenants]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string }
      responses:
        "201":
          description: Tenant created.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Tenant" }
        "409": { $ref: "#/components/responses/Conflict" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /api/v1/tenants/{id}:
    delete:
      operationId: deleteTenant
      summary: Delete a tenant (cascades to its keys)
      description: Requires the `platform` role.
      tags: [Tenants]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "204": { description: Deleted. }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/tenants/{id}/keys:
    get:
      operationId: listAPIKeys
      summary: List a tenant's API keys
      description: Requires `editor`. Per-tenant admins see only their own tenant.
      tags: [Keys]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: Keys (no plaintext).
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/APIKey" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
    post:
      operationId: createAPIKey
      summary: Mint an API key
      description: Requires `admin`. Plaintext is returned exactly once.
      tags: [Keys]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, role]
              properties:
                name: { type: string }
                role: { type: string, enum: [viewer, editor, admin] }
      responses:
        "201":
          description: Key minted.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/NewKeyResult" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /api/v1/tenants/{id}/keys/rotate:
    post:
      operationId: bulkRotateTenantKeys
      summary: Rotate every key in a tenant
      description: >
        All-or-nothing transactional rotation. `?except=self` spares the
        caller's own key. Requires `admin`.
      tags: [Keys]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
        - { name: except, in: query, schema: { type: string, enum: [self] } }
      responses:
        "200":
          description: Rotation results — each with a once-only plaintext.
          content:
            application/json:
              schema:
                type: object
                properties:
                  results:
                    type: array
                    items: { $ref: "#/components/schemas/NewKeyResult" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /api/v1/keys/{id}:
    patch:
      operationId: updateAPIKeyRole
      summary: Change an API key's role
      description: Requires `admin`. Emits `api_key.role_changed`.
      tags: [Keys]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [role]
              properties:
                role: { type: string, enum: [viewer, editor, admin] }
      responses:
        "200":
          description: Updated key.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/APIKey" }
        "403": { $ref: "#/components/responses/Forbidden" }
    delete:
      operationId: deleteAPIKey
      summary: Delete an API key
      description: Requires `admin`. Emits `api_key.deleted`.
      tags: [Keys]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "204": { description: Deleted. }
        "403": { $ref: "#/components/responses/Forbidden" }
  /api/v1/keys/{id}/rotate:
    post:
      operationId: rotateAPIKey
      summary: Rotate an API key's secret in place
      description: >
        Preserves id/name/role/tenant; only the secret changes. Flushes the
        auth cache and emits `api_key.rotated` with old + new prefixes. Requires
        `admin`.
      tags: [Keys]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: Rotated — plaintext shown once.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/NewKeyResult" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /api/v1/keys/me/rotate:
    post:
      operationId: rotateMyAPIKey
      summary: Rotate the calling key in place
      description: Self-service rotation. Any authenticated role (`viewer`).
      tags: [Keys]
      responses:
        "200":
          description: Rotated — plaintext shown once.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/NewKeyResult" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/v1/keys/me/burn:
    post:
      operationId: burnMyAPIKey
      summary: Rotate the calling key WITHOUT returning the new plaintext
      description: >
        Emergency response to a compromised session — the new secret is never
        returned, so the caller is locked out and must recover out-of-band.
      tags: [Keys]
      responses:
        "204": { description: Burned. }
        "401": { $ref: "#/components/responses/Unauthorized" }

  # ---------------------------------------------------------------------------
  # Protocol Endpoints
  # ---------------------------------------------------------------------------
...
components:
...
  schemas:
...
    Error:
      type: object
      description: >
        The management API's error envelope. Every 4xx and 5xx from a
        /api/v1/... route has this shape.

        NOT what the provider endpoints return when a quota rejects a request:
        those mimic the upstream `{"error": {"type", "message"}}` object so an
        SDK surfaces them correctly. See ProviderQuotaError.
      required: [error]
      properties:
        error:
          type: string
          example: agent "weather-agent" not found
        available_agents:
          type: array
          items: { type: string }
          description: >
            Present on a 404 for an unknown agent — what the caller could have
            asked for instead. Tenant-scoped: it never names another tenant's
            agents.
        details:
          type: string
          description: >
            Present when a write or reload failed schema validation; carries the
            validator's output.
...
    Tenant:
      type: object
      required: [id, name, created_at]
      properties:
        id: { type: string, example: ten_e19321097fc42a65 }
        name: { type: string, example: default }
        created_at: { type: string, format: date-time }
    APIKey:
      type: object
      description: Public view of an API key — never includes the secret.
      required: [id, name, prefix, role, created_at]
      properties:
        id: { type: string, example: key_a1b2 }
        name: { type: string, example: ci-bot }
        prefix: { type: string, example: mak_4d00115a, description: Public key prefix (no secret). }
        role: { type: string, enum: [viewer, editor, admin, platform] }
        tenant_id: { type: string }
        created_at: { type: string, format: date-time }
        last_used: { type: string, format: date-time, nullable: true }
    NewKeyResult:
      type: object
      description: Returned once on mint/rotate — carries the only copy of the plaintext.
      required: [key, plaintext]
      properties:
        key: { $ref: "#/components/schemas/APIKey" }
        plaintext:
          type: string
          example: mak_4d00115a_neV5hFQoDiw4ajfeQLvq6vLru6TQkBqz
          description: Shown exactly once; bcrypt-hashed immediately server-side.
...
  responses:
...
    NotFound:
      description: The requested resource was not found.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error:
              code: not_found
              message: The requested resource was not found.
              details: {}
...
    Unauthorized:
      description: Missing or invalid API key (multi-tenant mode).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error:
              message: missing Authorization bearer token or X-Api-Key header
              type: authentication_error
    Forbidden:
      description: Authenticated, but the key's role is below the route's floor.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Conflict:
      description: The resource already exists (e.g. duplicate tenant name).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"

  # ---------------------------------------------------------------------------
  # Security Schemes
  # ---------------------------------------------------------------------------
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: >
        `Authorization: Bearer <api-key>`. Required on `/api/v1/*` routes when
        the server runs with `MOCKAGENTS_MULTI_TENANT=1`.
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-Api-Key
      description: "Alternative to `Authorization: Bearer` for the same API key."

# Applies to the management API in multi-tenant mode. Protocol endpoints
# (/v1/*) are always unauthenticated and override this with `security: []`.
security:
  - BearerAuth: []
  - ApiKeyAuth: []